requirements.txt
alembic==1.15.2
APScheduler==3.11.0
Authlib==1.5.2
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.2
dicttoxml==1.7.16
Flask==3.1.0
Flask-APScheduler==1.13.1
flask-cors==6.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
flexcache==0.3
flexparser==0.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==2.1.0
Pint==0.24.4
platformdirs==4.3.8
pycountry==24.6.1
pycparser==2.22
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
toml==0.10.2
typing_extensions==4.13.1
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
xmltodict==0.14.2
psycopg2-binary

